mcp-server/pyproject.toml
[project]
name = "ai-brain-mcp"
version = "0.1.0"
description = "MCP server for persistent memory via mem0"
requires-python = ">=3.11"
dependencies = [
    "mem0ai>=0.1.30,<2.0",
    "mcp[cli]>=1.0",
    "pydantic-settings>=2.0",
]

[project.scripts]
ai-brain-mcp = "ai_brain_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_brain_mcp"]
